+++ src/app/admin/page.tsx
@@ -8,7 +8,8 @@ export default function Component() {
     { text: "Home", url: "/admin" },
     { text: "Pricing", url: "/pricing" },
     { text: "About", url: "/about-us" },
-    { text: "Login / Signup", url: "/api/auth/login" },
+    { text: "Sign Up", url: "/sign-in" },
+    { text: "Log In", url: "/log-in" },
   ];
   return (
     <div>
