+++ src/components/navbar.jsx
@@ -2,8 +2,11 @@
 
 import React from "react";
 import Link from "next/link";
+import { authClient } from "@/lib/auth-client";
+import { useRouter } from "next/navigation";
 
 function Navbar({ links }) {
+  const router = useRouter();
   return (
     <header className="flex h-16 flex-col items-center border-y px-4 sm:h-14 sm:flex-row lg:px-6">
       <Link className="flex items-center justify-center" href="/admin">
@@ -11,14 +14,31 @@ function Navbar({ links }) {
         <span className="ml-2 font-semibold">LMIA Opportunities</span>
       </Link>
       <nav className="flex items-center justify-center gap-4 py-2 sm:ml-auto sm:mt-0 sm:gap-6">
-        {links.map((link, index) => (
-          <Link
-            key={index}
-            className="text-sm font-medium underline-offset-4 hover:underline sm:text-base"
-            href={link.url}>
-            {link.text}
-          </Link>
-        ))}
+        {links.map((link, index) =>
+          link.text === "Log Out" ? (
+            <button
+              key={index}
+              className="text-sm font-medium underline-offset-4 hover:underline sm:text-base"
+              onClick={() =>
+                authClient.signOut({
+                  fetchOptions: {
+                    onSuccess: () => {
+                      router.push("/admin");
+                    },
+                  },
+                })
+              }>
+              Log Out
+            </button>
+          ) : (
+            <Link
+              key={index}
+              className="text-sm font-medium underline-offset-4 hover:underline sm:text-base"
+              href={link.url}>
+              {link.text}
+            </Link>
+          )
+        )}
       </nav>
     </header>
   );
